refactor: move api functions closer to the struct they belong to
