+++ api/kyverno/v1/clusterpolicy_types.go
@@ -1,6 +1,8 @@
 package v1
 
 import (
+	"strings"
+
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -25,6 +27,80 @@ type ClusterPolicy struct {
 	// Status contains policy runtime data.
 	// +optional
 	Status PolicyStatus `json:"status,omitempty" yaml:"status,omitempty"`
+}
+
+// HasAutoGenAnnotation checks if a policy has auto-gen annotation
+func (p *ClusterPolicy) HasAutoGenAnnotation() bool {
+	annotations := p.GetAnnotations()
+	val, ok := annotations["pod-policies.kyverno.io/autogen-controllers"]
+	if ok && strings.ToLower(val) != "none" {
+		return true
+	}
+
+	return false
+}
+
+// HasMutateOrValidateOrGenerate checks for rule types
+func (p *ClusterPolicy) HasMutateOrValidateOrGenerate() bool {
+	for _, rule := range p.Spec.Rules {
+		if rule.HasMutate() || rule.HasValidate() || rule.HasGenerate() {
+			return true
+		}
+	}
+	return false
+}
+
+// HasMutate checks for mutate rule types
+func (p *ClusterPolicy) HasMutate() bool {
+	for _, rule := range p.Spec.Rules {
+		if rule.HasMutate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasValidate checks for validate rule types
+func (p *ClusterPolicy) HasValidate() bool {
+	for _, rule := range p.Spec.Rules {
+		if rule.HasValidate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasGenerate checks for generate rule types
+func (p *ClusterPolicy) HasGenerate() bool {
+	for _, rule := range p.Spec.Rules {
+		if rule.HasGenerate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasVerifyImages checks for image verification rule types
+func (p *ClusterPolicy) HasVerifyImages() bool {
+	for _, rule := range p.Spec.Rules {
+		if rule.HasVerifyImages() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// BackgroundProcessingEnabled checks if background is set to true
+func (p *ClusterPolicy) BackgroundProcessingEnabled() bool {
+	if p.Spec.Background == nil {
+		return true
+	}
+
+	return *p.Spec.Background
 }
 
 // ClusterPolicyList is a list of ClusterPolicy instances.
